debug: Add drgn script
drgn can be a useful tool to dump some information in case of a hang in
md. Add as script which dumps some stats on active stripes and the
number of stripes in each state.

This requires a relatively recent version of drgn to be installed
(let's say 0.0.20 which isn't released yet).

There are instructions to install from source here:

  https://github.com/osandov/drgn

In the future, it may be installed with pip.