md: Don't try to figure out disk type from sysfs in _get_next_disk()
get_special_disk() is called in setup() before the array is setup, so
trying to read from sysfs to get the disk (or size) for the journal
does not work.

Change it to store the disk_type and disk_size and use that for
creating extra ram or loop disks.