md: Moved argument capture into MDInstance init
allocating special devices requires knowlege of the devices to be used
for testing and test3 attempts to allocate a verification device before
setting up the md instance in order to zero the devices before starting
the test. Fix this by getting arguments and storing them in MDInstance
init for later use by get_special_device() and setup().