test3: Remove md dev iterator from verification check
My previous test3 changes where tested on a system which had run many
tests previously, which missed the fact that we can't iterate throught
the md devices if it has not yet been created. Fortunatly we have a list
of devices to be used laying arround so just use that instead.