perf_test: fix peak detection
When searching for the best performance across all results, we sort the
results list by different metrics (for example 'Seq Write BW').

However, these are stored as strings with suffixes (such as 9.3 GiB/s).

This "works" as long as all metrics have the same suffix (such as GiB/s)
and that all have the same number of whole digits (for example, this
breaks when comparing 9.3 and 10.0, since '9.3' > '10.0').

To fix this, store the metric with and without the suffix, so that we
can sort by latter and print the former.

Fixes: 3a84e932d71d ("perf_test: add performance test script")