funccount: introduce funccount wrapper
funccount is a script from perf-tools[1] that uses ftrace to count how
many times kernel functions were called.

This python module is a tiny wrapper that makes it convenient to use and
returns a dictionary of all functions requested and their counts.

Note that in order to trace multiple functions, a fork of funccount is
required[2].

The optional sleep parameter will be useful when this is used with the
perf_tool, so that we can ignore IO done during the fio ramp_time.

[1]: https://github.com/brendangregg/perf-tools
[2]: https://github.com/brendangregg/perf-tools/pull/115