perf_test: add funccount support
Add a --funccount option to count how many IOs go through the slow/fast
paths.

Note that the slow path is only an approximation, as there isn't a
function that is always (and only) called by the slow path.

This is only available for the write path, as all functions that we
could potentially use for the fast read path were inlined.