perf_test: add --rebuild option
This option runs the test during array rebuild.

While it would be possible to set it into rebuild mode once and run
multiple tests, we wouldn't know when/if the rebuild has finished and if
we're back to non-degraded mode.

Thus, always force a rebuild before each test. This requires us to be
able to stop the arrays without the wait (which could take hours
depending on their size).