debug: Add drgn script for some examples to tracing a raid5 hang
This script helped debug one of the recent bug's but it was pretty
much specific to that bug. It is a good example to use for future
issues.

The code dumps bios stuck in blk-plugs as well as bios stuck in
tasks hung in the block layer writeback throttling system.